Webpage was offline
I just saw our webpage was offline, for some reason, probably linked to the custom domain we added. I don't know for how long it has been offline, obviously it was working fine when we handed the work in on deadline day. A link to a copy of the website was added to the readme in case this issue arises again. You can verify our last website modification was done before the deadline at https://nicovp1.github.io/ada_website/ .